Implemented secret manager for api key
Rather than storing the api key as an environment variable, we will instead have an environment variable containing a string pointing to a secret, which will be called at runtime. That way, we can avoid hardcoding the api key, while also not having to keep the entire ARN in the code.